Wipe fs.c and rename lowercase logger
